+++ src/main/kotlin/us/frollo/frollosdk/database/SDKDatabase.kt
@@ -630,7 +630,8 @@ abstract class SDKDatabase : RoomDatabase() {
                 // New changes in this migration:
                 // 1) Alter table - provider - add joint_accounts_available & associated_provider_ids columns
                 // 2) Alter account table - Delete products_available column
-                // 3) Drop & re-create table `service_outage`
+                // 3) Drop & re-create table service_outage
+                // 4) Drop & re-create table addresses
 
                 database.execSQL("ALTER TABLE `provider` ADD COLUMN `joint_accounts_available` INTEGER")
                 database.execSQL("ALTER TABLE `provider` ADD COLUMN `associated_provider_ids` TEXT")
@@ -648,8 +649,21 @@ abstract class SDKDatabase : RoomDatabase() {
                 database.execSQL("COMMIT")
                 // END - Drop column products_available
 
+                // START - Drop & re-create table service_outage
+                database.execSQL("BEGIN TRANSACTION")
                 database.execSQL("DROP TABLE service_outage")
                 database.execSQL("CREATE TABLE IF NOT EXISTS `service_outage` (`type` TEXT NOT NULL, `start_date` TEXT NOT NULL, `end_date` TEXT, `duration` INTEGER, `outage_id` INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, `read` INTEGER NOT NULL, `message_title` TEXT NOT NULL, `message_summary` TEXT NOT NULL, `message_description` TEXT NOT NULL, `message_action` TEXT NOT NULL, `message_url` TEXT)")
+                database.execSQL("COMMIT")
+                // END - Drop & re-create table service_outage
+
+                // START - Drop & re-create table addresses
+                database.execSQL("BEGIN TRANSACTION")
+                database.execSQL("DROP INDEX IF EXISTS `index_addresses_address_id`")
+                database.execSQL("DROP TABLE addresses")
+                database.execSQL("CREATE TABLE IF NOT EXISTS `addresses` (`address_id` INTEGER NOT NULL, `dpid` TEXT, `line_1` TEXT, `line_2` TEXT, `suburb` TEXT, `town` TEXT, `region` TEXT, `state` TEXT, `country` TEXT, `postal_code` TEXT, `long_form` TEXT, PRIMARY KEY(`address_id`))")
+                database.execSQL("CREATE INDEX IF NOT EXISTS `index_addresses_address_id` ON `addresses` (`address_id`)")
+                database.execSQL("COMMIT")
+                // END - Drop & re-create table addresses
             }
         }
     }